wip: factoring out common code
